HSA-84: reorganized all dockerfiles into a folder called Deployment. Added new azure job for integration tests. Created Dockerfile for integration.

+++ HSA-backend/features/environment.py
@@ -1,5 +1,7 @@
 import os
 from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.chrome.service import Service
 import sys
 import subprocess
 import signal
@@ -9,6 +11,7 @@ import traceback
 from behave import fixture, use_fixture
 from django.core.management import call_command
 from django.core.management.base import CommandError
+import os
 
 def block_for_server(url):
      # Wait for Angular to be ready
@@ -54,16 +57,33 @@ def before_scenario(context, scenario):
         print(f'Error flushing the database: {e}')
 
 def before_all(context):
+    for evn in os.environ:
+        print (evn,"=", os.environ[evn])
     try:
         path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../..'))
-        sys.path.append(path) # we need this so python can find our app as a module
+        sys.path.append(path)  # we need this so python can find our app as a module
         os.environ["DJANGO_SETTINGS_MODULE"] = "hsabackend.settings"
-        context.url = "http://localhost:4200"
-        path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../../HSA-frontend'))
-        os.chdir(path)
-        context.angular = subprocess.Popen(["ng", "serve"])
-        block_for_server("http://localhost:4200")
-        context.browser = webdriver.Chrome()
+        if "INTEGRATION_FLAG" in os.environ:
+            chrome_options = webdriver.ChromeOptions()
+            chrome_options.add_argument('--headless')  # Run in headless mode
+            chrome_options.add_argument('--disable-gpu')  # Disable GPU usage (often recommended)
+            chrome_options.add_argument('--window-size=1920,1080')  # Set a binary size (if needed)
+            chrome_options.add_argument('--no-sandbox')  # (Optional) May help in some CI environments
+            chrome_options.add_argument('--disable-dev-shm-usage')  # (Optional) Overcome limited resource problems
+            context.url = "http://localhost:8000"
+            try:
+                context.browser = webdriver.Chrome(options=chrome_options)
+            except Exception as e:
+                print(f"Error creating WebDriver: {e}")
+                raise
+        else:
+            print("CONTEXT IGNORED, Integration flag value was set to: ", os.environ["INTEGRATION_FLAG"])
+            context.url = "http://localhost:4200"
+            path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../../HSA-frontend'))
+            os.chdir(path)
+            context.angular = subprocess.Popen(["ng", "serve"])
+            block_for_server("http://localhost:4200")
+            context.browser = webdriver.Chrome()
         use_fixture(django_server, context)  # Start server before all scenarios
         
     except Exception as e:
@@ -73,5 +93,8 @@ def before_all(context):
         
 
 def after_all(context):
-    context.angular.send_signal(signal.SIGINT)
-    context.angular.wait()
+    if "INTEGRATION_FLAG" in os.environ:
+        context.browser.quit()
+    else:
+        context.angular.send_signal(signal.SIGINT)
+        context.angular.wait()
